Upgrade version to 1.52. Upgrade logging library versions. kotlin to 1.20. Put SchnorrProof back in common.proto. Cleanup compile messages.
